feat: alter IModelService implementation to only use generic interface
Change the implementation for model services to allow for using a generic interface.
This replaces the use of model-specific interfaces built to obscure the parent key type.
Instead, the parent key of batch model services is defined as a generic object type. The
implementation will be responsible for checking the type and handling mismatches.