feat: move recurring elements to custom layout
Moves the page header defintion to a page layout, but updates cause improper
headers to be applied to pages. Since a new component is initialized for a different
model, but the ViewService instance is the same, updating the header updates it for
all open pages with the same layout applied.