Feat/authentication (#10)
* feat: add default identity scaffolding for ASP.NET Core

* add default token providers

* fix: remove all identity scaffolding to resolve page routing issue

* break: re-introduce routing error

Routing error appears related to multiple Login pages being defined.

* temporarily remove authentication service registration

* fix: resolve authentication scaffolding issues

Replaces EulerFinancial.Blazor with EulerFinancial.Web to take advantage of new 6.0
features and result issues with identity scaffolding.

* chore: remove obsolete project files

* fix: separate navigation component from localizable component

* fix: resolve data context saving issues related to service lifetime

Fixes issues with saving data across pages caused by the service lifetime of
DbContext instances created with depenendency injection. Changes the service
class to use IDbContextFactory and leave the dispose to the model service.

* feat: add logic for disabling components or rendering as readonly

* feat: move recurring elements to custom layout

Moves the page header defintion to a page layout, but updates cause improper
headers to be applied to pages. Since a new component is initialized for a different
model, but the ViewService instance is the same, updating the header updates it for
all open pages with the same layout applied.

* fix: fix page title formatting for home page