fix(DAL): add fix for updating InvestmentStrategy entity graph
When saving InvestmentStrategy as an entity graph, the attempt fails due to a
discrepancy between the observed and expected affected row count. Add a function
to the InvestmentStrategyService to perform the complex delete and update for child
records of the Investment Strategy entity.

Additonal context: https://docs.microsoft.com/en-us/ef/core/saving/disconnected-entities

Resolves #80