feat(Blazor): finish example implementation of BankTransactionCode attribute entry editor
Adds test implementation of editors for attributes applied to BankTransactionCode instances. Approach is similar to that of InvestmentStrategy, but differs in that each group expects only one entry. Error is raised when trying to add new group, and presentation is not fitting the model. Needs to be re-designed to new form.