feat(DAL): review view model implementation to include unweighted attribute entries
Add abstract class for handling management of attribute entry collections that always have a weight of 100% for a given date.
Rename the view model class names and interfaces to be more descriptive of the purpose of those objects.
Define missing ModelDisplay entries and update the ModelNoun values for attribute entries.