fix(Blazor): resolve formatting issues with modal forms
Resolve issue where modal form button hover/click behavior was not working in some window sizes. Root cause was the z-index applied due to the position of the modal form definitions in the page, resulting in positioning behind the page body content. Move modal definitions to the end of the razor syntax section to avoid this issue.