Refactor/reference data service (#93)
* refactor(DAL): improve ReferenceDataService with generic helper classes

Simplify ReferenceDataService by consolidating behavior in IQueryBuilder and IQueryDataStore generic interace. Interfaces are responsible for handling query creation and execution, with IReferenceDataService wrapping these interfaces with simple built-in queries.

* refactor(Blazor): simplify calls for DTOs using IReferenceDataService

Also eliminates much of the need for ToLookups extension methods.

* refactor(Blazor): move method for return display value for given DTO key to extension method

* refactor(DAL): revise ReferenceDataService to include DTO selects

Add DTO selection methods to simplify accessing models and converting them into DTOs for
specifying foreign key values in write operations.

* test(DAL): add simple tests for built-in Reference Data queries

* fix(Blazor): adapt UI components to changes in reference data features

Resolves errors resulting from breaking changes to IReferenceDataService and LookupModel. Uses more
generalized functions in IReferenceDataService with options for built-in methods for common queries, as well.

* style(DAL): simplify 'new' expression

* fix(DAL): change IReferenceDataService to singleton and add comments

* refactor: rename IReferenceDataService to IQueryService and change namespace

Move renamed IQueryService to ~ModelService.Query namespace, as it is more in keeping the with the
features of the interfaces in that location. Access will be restricted once wrapped in Contoller methods.

* refactor: merge IQueryService behavior into IQueryController

Replace IReferenceDataService with IQueryService and wraps service in controller. IQueryController
is a property of IController and IBatchController interfaces. Narrows the exposure area of the data access
layer.