Feat/issue#119 (#127)
* feat(Blazor): add attributes to Account detail and edit pages

Add view and edit support for attributes on the Account detail and edit pages.

Resolves #119

* fix(Blazor): correct error in retrieving ModelAttribute lookup values

Modify the function 'GetAttributeMembers' to be a synchronous wrapper of an async method
that makes a call via the IQueryService instance. Handles all lookup lists better by avoiding
dependency on defined attributes of the parent entity.