Feat: Adds support for building project to Docker container (#142)
* fix: minor fixes to project stucture and configuration

Fix issue with program configuration builder that caused a
crach when run in a Linux OS.
Minor fix to the path to the NjordFinance.Web.Test project.
Remove option to emit compiler-generated files causing a build
failure in NjordFinance.Web.Areas.Identity.

* fix: adjust context definitions to use in memory by default

Change context classes to default to in-memory databases if
options are not configured, instead of SQL Server. Done to avoid
references to connection strings that do not exist without SQL
Server.

* build(NjordFinance.Web): add Dockerfile and launch profile

Add Dockerfile for use in buiding and deploying app in a Docker
container. Use in conjunction with the docker build command to map ports from the container to the local machine.

Related #141

* fix(EntityModel): fix error in EntityModel.FinanceDbContext

Remove duplicated OnConfiguring method. See definition in file
FinanceDbContextExt.cs.

* update README.md