feat(Blazor): add service responsible for handling HTTP requests for Razor
Isolates the HTTP request/response management in a service class to reduce code
repetition in model pages. Currently limited to controller actions for POST/PUT on
a single resource at a time, and GET actions for retrieving a single resource or a set
of resources of a given type.

Related #214