feat: Add basic HTTP service for Razor components (#215)
feat(Blazor): add service for managing HTTP requests from the UI layer
Isolates the HTTP request/response management in a service class to reduce code
repetition in model pages. Currently limited to controller actions for POST/PUT on
a single resource at a time, and GET actions for retrieving a single resource or a set
of resources of a given type.

Resolves #214

refactor(API): revise API to include a POST method to replace the GET with body method

RESTful APIs do not accept body content in GET methods. Add a workaround POST
method to [controller]/search endpoints, so that serialized 'ParameterDto' instances
can be passed as body content.