fix(DAL): convert certain entity model services to collection services
Some collections will be worked as collections instead and some further as
child objects to a parent record of another type. Fix the services enumerated in
#222 to inherit from ModelCollectionService<> instead of ModelService<>.

BREAKING CHANGE: Leaves the proper assignment of parent records to the UI
layer. API contollers perform the validation of the parent record assignment
where applicable.