refactor: large refactor merger targetting reference/query service (#231)
* fix(Blazor): partial work on clean-up of entity references in UI layer

* build(API): remove unnecessary package reference

* feat(API): add back support for posting searches to collection endpoints

* feat(Blazor): add collection methods to HttpService implementation

Revises the HttpService<> class to support the differing HTTP service  interfaces
based on whether the resource is worked as a collection or as an item.

* revise underlying services and update Create page references to entities

* additional clean-up replacing entity class references

- add endpoints to reference data controller
- clean-up page references to entity classes
- remove controller interfaces and classes from Web project
- resolve misc. issues with interfaces definitions

* apply metadata to DTO classes and misc. UI bug fixes

* refactor: replace reference data service with expanded query service methods

Replace IReferenceDataService with IQueryService methods to to avoid HTTP
requests that could be large and originate from the same server.

* fix: fix error mapping tuple return variable to enumerable collection

* test: update tests to include assignment to attribute DTO member

Fix tests to include assignment of value of AttributeMemberId member. Member
is a new addition to attribute entry records with aims to simplify DTO model.

* refactor(DAL): remove obsolete interface members and implementations

Remove obsolete DTO selection methods and replace with calls to methods that
supported paginated queries.

* refactor(Blazor): remove unnecessary nullable annotations

BREAKING CHANGE: removes reference data service support in favor of using IBuiltInQuery interface.